Bump Django to 3.2 LTS
Also bump other requirements that require Django to be >3

Pinned `django-munigeo` version to a git commit as no new version supporting Django 3 is released yet

Add `drf-oidc-auth` to requirements.in, as it is required for authentication and not a mandatory requirement by helusers. Version `<1` must be used as otherwise authentication will break.

Removed `modeltranslation` from `INSTALLED_APPS`, it hasn't been used for a long time