fix: prevent django from generating unnecessary migrations
If the value of PROTECTED_ROOT env variable
changes, Django would generate a migration
for this. However, local and production
environments might have different values
for that, which leads to warnings when
checking for missing migrations.

This commit adds a dummy storage subclass
to prevent Django from creating these
migrations.