feat: create images/files by reference, remove copy logic
Some of the serializers worked with a update-or-copy
logic instead of update-or-create (as their name would
imply), which was added in PR #363 (6c972342).

This caused an error whenever you're creating
a new hearing and have already uploaded a file to a
section (via the form's file upload functionality).
Added a test to confirm that this scenario functions
correctly now.

This removes that copy (or rather, save as copy)
logic and moves the burden to the consumer of the
API, i.e. the consumer needs to remove the IDs
from the data they want to copy.

However, files and images still need a reference
to the instance they're copied from, so add a
new field "reference_id" for this. This allows
the user to create new files or images based
on existing instances.

In its current state, it's meant for
"save as new" type of situations.
This means that if the pk is set for an
image/file, then it will ignore the
reference_id, i.e. you cannot update an
existing instance using reference_id.